feat:優化api    - POST /followships      若重複追蹤，回400，已經追蹤過
   - DELETE /followships/:id
     若沒有該帳號，回400，帳號不存在
     若沒有追蹤紀錄，回400，尚未追蹤

   - GET /users/:id/followings(followers)
     依照追蹤成立時間排序，越近靠前

   - GET /users/:id/likes
     依照按讚成立時間排序，越近靠前

   - GET /admin/users
     提供排序方式選擇 ex. /admin/users?order=tweet(以貼文數量排序，越多靠前)
     不填寫order，預設依照追蹤人數排序，越多靠前

fix: GET /users/:id/tweets，liked判斷有誤
   - 應根據當前登入者做判斷，而非以查詢對象帳號判斷

+++ controllers/followshipController.js
@@ -9,24 +9,35 @@ const makeFollow = async (req, res) => {
   if (!followingUser) {
     return res.status(400).json({ message: '追蹤失敗，該帳號不存在' })
   }
-  // TODO 無法追蹤兩次
   const user = helpers.getUser(req)
-  await Followship.create({
-    followerId: user.id,
-    followingId: req.body.id,
+  const [followship, created] = await Followship.findOrCreate({
+    where: {
+      followerId: user.id,
+      followingId: req.body.id,
+    },
   })
-  res.sendStatus(200)
+  if (created) {
+    return res.sendStatus(200)
+  }
+  return res.status(400).json({ message: '已經追蹤過' })
 }
 
-const unFollow = (req, res) => {
+const unFollow = async (req, res) => {
+  const followingUser = await User.findByPk(req.params.following_id)
+  if (!followingUser) {
+    return res.status(400).json({ message: '帳號不存在' })
+  }
   const user = helpers.getUser(req)
-  Followship.destroy({
+  const delCount = await Followship.destroy({
     where: {
       followerId: user.id,
       followingId: req.params.following_id,
     },
   })
-  res.sendStatus(200)
+  if (delCount === 1) {
+    return res.sendStatus(200)
+  }
+  return res.status(400).json({ message: '尚未追蹤' })
 }
 
 module.exports = { makeFollow, unFollow }
